Autodocs are generated for PDMesh using Documenter.jl.

PDMesh.Cone — Type

Cone(radius::AbstractFloat, length::AbstractFloat)

Cone shape.

Fields

radius::AbstractFloat: Radius of the cone.

length::AbstractFloat: Length of the cone.

Example

using PDMesh

# Create a cone
cone = Cone(1.0, 1.0)

# Create a mesh
mesh = create(cone; resolution=0.1, rand_=0.01, type=1)

See also

StandardCone()

Cone()

create

References

Wikipedia

PDMesh.Cone — Method

Cone()

Cone shape. Radius and length are 1.0.

see also

Cone

PDMesh.Cuboid — Type

Cuboid(bounds::Array{Float64, 2})

Cuboid shape.

Fields

bounds::Array{Float64, 2}: Bounds of the cuboid.

Example

using PDMesh

# Create a cuboid
cuboid = Cuboid([0.0 1.0; 0.0 1.0; 0.0 1.0])

# Create a mesh
mesh = create(cuboid)

See also

create

References

Wikipedia

PDMesh.Cuboid — Method

Cuboid()

Cuboid shape. Bounds are [-0.5 0.5; -0.5 0.5; -0.5 0.5]. Length is 1.0.

see also

create

StandardCuboid

PDMesh.Cylinder — Type

Cylinder

Cylinder shape.

Fields

radius::AbstractFloat: Radius of the cylinder.

thickness::AbstractFloat: Thickness of the cylinder.

length::AbstractFloat: Length of the cylinder.

Example

using PDMesh

# Create a cylinder
cylinder = Cylinder(1.0, 0.1, 2.0)

# Create a mesh
mesh = create(cylinder, resolution=0.1)

PDMesh.Disk — Type

Disk

Disk shape.

Fields

radius::AbstractFloat: Radius of the disk.

thickness::AbstractFloat: Thickness of the disk.

Example

using PDMesh

# Create a disk
disk = Disk(1.0, 0.1)

# Create a mesh
mesh = create(disk, resolution=0.1)

PDMesh.PostOpObj — Type

PostOpObj

These objects are used to create a post operation object (lazy) from object ( or post operation object)and an operation. Operations will be applied while a create call.

Fields

name::String: Name of the object.

objs::Vector{Any}: Objects to be operated on.

operations::Array{Function}: Operations to be applied.

PDMesh.Shape — Type

Shape

Abstract type for shapes.

PDMesh.Shell — Type

Shell

Shell shape.

Fields

outer_radius::AbstractFloat: Outer radius of the shell.

inner_radius::AbstractFloat: Inner radius of the shell.

Example

using PDMesh

# Create a shell
shell = Shell(1.0, 0.5)

# Create a mesh
mesh = create(shell, resolution=0.1)

PDMesh.Sphere — Type

Sphere

Sphere shape.

Fields

radius::AbstractFloat: Radius of the sphere.

Example

using PDMesh

# Create a sphere
sphere = Sphere(1.0)

# Create a mesh
mesh = create(sphere, resolution=0.1)

PDMesh.Cube — Method

Cube(L)

Cube shape. A special case of Cuboid.

Arguments

L::AbstractFloat: Length of the cube.

Example

using PDMesh

# Create a cube
cube = Cube(1.0)

# Create a mesh
mesh = create(cube)

See also

create

Cuboid

PDMesh.Cube — Method

Cube()

Cube shape. A special case of Cuboid. Length is 1.0.

see also

create

StandardCuboid

PDMesh.Indentor — Method

Indentor(angle, height; sides=3)

Indentor shape. A special case of Pyramid.

Arguments

angle::AbstractFloat: Apex angle of the Indentor.

height::AbstractFloat: Height of the Indentor.

Keyword Arguments

sides::Int: Number of sides of the Indentor.

Example

using PDMesh

# Create a Indentor
Indentor = Indentor(60.0, 1.0)

# Create a mesh
mesh = create(Indentor)

See also

create

PDMesh.Pyramid — Method

Pyramid

Pyramid shape.

Arguments

angle::AbstractFloat: Apex angle of the Pyramid.

Example

using PDMesh

# Create a Pyramid
Pyramid = Pyramid(60.0)

# Create a mesh
mesh = create(Pyramid)

See also

create

References

Wikipedia

PDMesh.StandardCone — Method

StandardCone()

Standard cone shape. A special case of Cone. Radius and length are 1.0.

Example

using PDMesh

# Create a standard cone
cone = StandardCone()

# Create a mesh
mesh = create(cone)

See also

Cone

PDMesh.StandardCuboid — Method

StandardCuboid()

Standard cuboid shape. A special case of Cuboid. Bounds are [-0.5 0.5; -0.5 0.5; -0.5 0.5]. Length is 1.0.

Example

using PDMesh

# Create a standard cuboid
cuboid = StandardCuboid()

# Create a mesh
mesh = create(cuboid)

See also

create

Cuboid

PDMesh.StandardCylinder — Method

StandardCylinder()

Standard cylinder shape. A special case of Cylinder. Radius is 1.0. Thickness is 0.3. Length is 1.0.

Example

using PDMesh

# Create a standard cylinder
cylinder = StandardCylinder()

# Create a mesh
mesh = create(cylinder)

See also

create

PDMesh.StandardDisk — Method

StandardDisk()

Standard disk shape. A special case of Disk. Radius is 1.0. Thickness is 0.3.

Example

using PDMesh

# Create a standard disk
disk = StandardDisk()

# Create a mesh
mesh = create(disk)

See also

create

PDMesh.StandardSphere — Method

StandardSphere()

Standard sphere shape. A special case of Sphere. Radius is 1.0.

Example

using PDMesh

# Create a standard sphere
sphere = StandardSphere()

# Create a mesh
mesh = create(sphere, resolution=0.1)

See also

create

PDMesh.changetype — Method

changetype(type::Array{Int64,1}, mask, ntype::Int64)

Change mesh particle type for object using boolean array mask.

Arguments

type::Array{Int64,1}: Array of particle types.

mask: Boolean array.

ntype::Int64: New particle type.

Returns

type::Array{Int64,1}: Array of particle types.

See also

changetype(obj::T, f::Function, ntype::Int64) where T <: SuperShape

PDMesh.changetype — Method

changetype(obj::T, f::Function, ntype::Int64) where T <: SuperShape

Change mesh particle type for object using function f.

Arguments

obj::T: Object.

f::Function: Function that returns a boolean array.

ntype::Int64: New particle type.

Returns

obj::T: Object.

See also

changetype(type::Array{Int64,1}, mask, ntype::Int64)

Example

using PDMesh

# Create a disk
disk = Disk(1.0, 0.1)

# Create a mesh
mesh = create(disk, resolution=0.1)

# Change particle type
mesh = changetype(mesh, out -> out['x'][:, 1] .> 0.0, 2)

PDMesh.combine — Method

combine(obj1::T1, obj2::T2) where {T1<:SuperShape, T2<:SuperShape}

Combine two objects. Duplicate of add.

PDMesh.create — Method

create(c::Cone; resolution=nothing, rand_=0.0, type::Int=1)

Create a mesh of a cone.

Arguments

c::Cone: Cone shape.

resolution=nothing: Resolution of the mesh.

rand_=0.0: Randomization factor.

type::Int=1: Type of the mesh.

Returns

Dict: Mesh.

Example

using PDMesh

# Create a cone
cone = Cone(1.0, 1.0)

# Create a mesh
mesh = create(cone)

See also

Cone

PDMesh.create — Method

create(c::Cuboid; resolution=nothing, rand_=0.0, type::Int=1)

Create a mesh of a cuboid.

Arguments

c::Cuboid: Cuboid shape.

resolution=nothing: Resolution of the mesh.

rand_=0.0: Randomization factor.

type::Int=1: Type of the mesh.

Returns

Dict: Mesh.

Example

using PDMesh

# Create a cuboid
cuboid = Cuboid([0.0 1.0; 0.0 1.0; 0.0 1.0])

# Create a mesh
mesh = create(cuboid)

See also

Cuboid

PDMesh.create — Method

create(c::Cylinder; resolution=nothing, rand_=0.0, type::Int=1)

Create a mesh from a cylinder.

Arguments

c::Cylinder: Cylinder object.

resolution=nothing: Resolution of the mesh.

rand_=0.0: Randomization factor.

type::Int=1: Type of the mesh.

Returns

out::Dict{Symbol, Any}: Dictionary containing the mesh data.

Example

using PDMesh

# Create a cylinder
cylinder = Cylinder(1.0, 0.1, 2.0)

# Create a mesh
mesh = create(cylinder, resolution=0.1)

PDMesh.create — Method

create(c::Disk; resolution=nothing, rand_=0.0, type::Int=1)

Create a mesh from a disk.

Arguments

c::Disk: Disk object.

resolution=nothing: Resolution of the mesh.

rand_=0.0: Randomization factor.

type::Int=1: Type of the mesh.

Returns

out::Dict{Symbol, Any}: Dictionary containing the mesh data.

Example

using PDMesh

# Create a disk
disk = Disk(1.0, 0.1)

# Create a mesh
mesh = create(disk, resolution=0.1)

PDMesh.create — Method

create(pobj::PostOpObj, args...; kwargs...)

Create a mesh from a post operation object.

Arguments

pobj::PostOpObj: Post operation object.

args...: Arguments to be passed to the create function.

kwargs...: Keyword arguments to be passed to the create function.

Returns

out::Dict{Symbol, Any}: Dictionary containing the mesh data.

PDMesh.create — Method

create(s::Sphere; resolution=nothing, rand_=0.0, type::Int=1)

Create a mesh from a sphere or shell.

Arguments

s::Sphere: Sphere object.

resolution=nothing: Resolution of the mesh.

rand_=0.0: Randomization factor.

type::Int=1: Type of the mesh.

Returns

out::Dict{Symbol, Any}: Dictionary containing the mesh data.

PDMesh.create — Method

create(shape::T; resolution=nothing, rand_=0.0, type::Int=1) where T <: Shape

Abstact function for creating Shape objects.

Arguments

shape::T: Shape object.

resolution=nothing: Resolution of the mesh.

rand_=0.0: Randomization factor.

type::Int=1: Type of the mesh.

Returns

out::Dict{Symbol, Any}: Dictionary containing the mesh data.

PDMesh.delete — Method

delete(obj::T, f::Function) where T <: SuperShape

Delete mesh particles for object using function f.

Arguments

obj::T: Object.

f::Function: Function that returns a boolean array.

Returns

obj::T: Object.

Example

using PDMesh

# Create a disk
disk = Disk(1.0, 0.1)

# Create a mesh
mesh = create(disk, resolution=0.1)

# Delete particles
mesh = delete(mesh, out -> out['x'][:, 1] .> 0.0)

See also

delete(out, mask::BitArray)

PDMesh.make — Method

make(shape::T) where T <: Shape

Create a mesh from a shape. This function is a wrapper for the create function. It is used to create a mesh from a shape with default arguments.

Arguments

shape::T: Shape object.

Returns

out::Dict{Symbol, Any}: Dictionary containing the mesh data.

PDMesh.move — Method

move(x::Array{Float64,2}; by=[0.0, 0.0, 0.0])

Move mesh particles for object by given "by".

Arguments

x::Array{Float64,2}: Array of mesh particles.

by=[0.0, 0.0, 0.0]: Translation vector.

Returns

x::Array{Float64,2}: Array of mesh particles.

See also

move

PDMesh.move — Method

move(obj::T; by=[0.0, 0.0, 0.0]) where T

Move mesh particles for object by given "by".

Arguments

obj::T: Object.

by=[0.0, 0.0, 0.0]: Translation vector.

Returns

obj::T: Object.

Example

using PDMesh

# Create a disk
disk = Disk(1.0, 0.1)

# Create a mesh
mesh = create(disk, resolution=0.1)

# Move particles
mesh = move(mesh, by=[0.0, 0.0, 0.1])

See also

move

PDMesh.repack! — Method

repack!(d::Dict, keys_, vals)

Repack a dictionary from its components inplace.

Arguments

d::Dict: Dictionary to be repacked.

keys_: Keys of the dictionary.

vals: Values of the dictionary.

Returns

d::Dict: Dictionary containing the components.

PDMesh.repack — Method

repack(args...; keys_ = (:x, :v, :y, :volume, :type))

Repack a dictionary from its components.

Arguments

args...: Components to be packed.

keys_ = (:x, :v, :y, :volume, :type): Keys of the dictionary.

Returns

d::Dict: Dictionary containing the components.

PDMesh.rotate — Method

rotate(x::Array{Float64,2}; angle=0.0, point=[0.0, 0.0, 0.0], vector_=[1.0, 0.0, 0.0])

Rotate mesh particles for object by given angle about given vector and point.

Arguments

x::Array{Float64,2}: Array of mesh particles.

angle=0.0: Rotation angle.

point=[0.0, 0.0, 0.0]: Rotation point.

vector_=[1.0, 0.0, 0.0]: Rotation vector.

Returns

x::Array{Float64,2}: Array of mesh particles.

See also

rotate

PDMesh.rotate — Method

rotate(obj::T; angle=0.0, point=[0.0, 0.0, 0.0], vector_=[1.0, 0.0, 0.0]) where T <: SuperShape

Rotate shape object by given angle about given vector and point.

PDMesh.unpack — Method

unpack(d::Dict)

Unpack a dictionary into its components.

Arguments

d::Dict: Dictionary to be unpacked.

Returns

x::Array{Float64, 1}: x coordinates.

v::Array{Float64, 1}: v coordinates.

y::Array{Float64, 1}: y coordinates.

volume::Array{Float64, 1}: Volume of the mesh.

type::Array{Int64, 1}: Type of the mesh.

PDMesh.velocity — Method

velocity(v::Array{Float64,2}; velocity=[0.0, 0.0, 0.0])

Change velocity of particles for object by given "velocity".

Arguments

v::Array{Float64,2}: Array of mesh velocities.

velocity=[0.0, 0.0, 0.0]: Velocity vector.

Returns

v::Array{Float64,2}: Array of mesh velocities.

PDMesh.velocity — Method

velocity(v::Array{Float64,2}; velocity=[0.0, 0.0, 0.0])

Change velocity of particles for object by given "velocity".

Arguments

v::Array{Float64,2}: Array of mesh velocities.

velocity=[0.0, 0.0, 0.0]: Velocity vector.

Returns

v::Array{Float64,2}: Array of mesh velocities.

PDMesh.velocity — Method

velocity(obj::T, f::Function, vel) where T <: SuperShape

Change velocity of particles for object using boolean array from function f.

Arguments

obj::T: Object.

f::Function: Function that returns boolean array.

vel: Velocity vector.

Returns

obj::T: Object.

PDMesh.write_data — Method

write_data(filename::String, out::Dict)

Write data to a file in the following format:

    N
    # id, type, position, velocity, volume
    1, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
    2, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
    3, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
    4, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
    5, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0

where N is the number of particles, id is the particle id, type is the particle type,
position is the particle position, velocity is the particle velocity, and
volume is the particle volume.

Arguments

filename::String: Name of the file to write to.

out::Dict: Dictionary containing the following keys:

:x: Position of the particles.

:v: Velocity of the particles.

:type: Type of the particles.

:volume: Volume of the particles.

Example

using PDMesh

# Create a cube
cube = Cube(1.0, 1.0, 1.0)

# Create a dictionary of the union
out = create(cube)

# Write the data to a file
write_data("data.txt", out)

See also

write_data_peridigm

write_data

PDMesh.write_data — Method

write_data(filename::String, x::Matrix, v::Matrix, type::Vector,  vol::Vector)

Write data to a file in the following format:

N
# id, type, position, velocity, volume
1, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
2, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
3, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
4, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
5, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0

where N is the number of particles, id is the particle id, type is the particle type, position is the particle position, velocity is the particle velocity, and volume is the particle volume.

Arguments

filename::String: Name of the file to write to.

x::Matrix: Position of the particles.

v::Matrix: Velocity of the particles.

type::Vector: Type of the particles.

vol::Vector: Volume of the particles.

see also

write_data

PDMesh.write_data — Method

write_data(filename::String, obj::SuperShape; kwargs...)

Write data to a file in the following format:

    N
    # id, type, position, velocity, volume
    1, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
    2, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
    3, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
    4, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0
    5, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0

where N is the number of particles, id is the particle id, type is the particle type,
position is the particle position, velocity is the particle velocity, and
volume is the particle volume.

Arguments

filename::String: Name of the file to write to.

obj::SuperShape: Shape or post operation object to write to file.

Keyword Arguments

kwargs...: Keyword arguments to pass to create.

Example

using PDMesh

# Create a cube
cube = Cube(1.0, 1.0, 1.0)

# Write the data to a file
write_data("data.txt", cube)

See also

write_data_peridigm

write_data

PDMesh.write_data_peridigm — Method

write_data_peridigm(filename::String, x::Matrix, type::Vector,  vol::Vector)

Write data to a file in the following format:

    N
    # position, type, volume
    0.0, 0.0, 0.0, 1, 1.0
    0.0, 0.0, 0.0, 1, 1.0
    0.0, 0.0, 0.0, 1, 1.0
    0.0, 0.0, 0.0, 1, 1.0
    0.0, 0.0, 0.0, 1, 1.0

where N is the number of particles, type is the particle type, position is the particle position, and volume is the particle volume.

Arguments

filename::String: Name of the file to write to.

x::Matrix: Position of the particles.

type::Vector: Type of the particles.

vol::Vector: Volume of the particles.

see also

write_data