Autodocs

PDMesh.Cone — Type

Cone

Shape object of type cone.

Args

- radius : radius of cone
- length : length of cone

Returns

- obj : Shape object **Cone**.

PDMesh.Cuboid — Type

Cuboid

Shape object of type cuboid.

Args

- bounds : min, max of cuboid for e.g. [-10 10; 0 3; -2 2]

Returns

- obj : Shape object **Cuboid**.

PDMesh.Cylinder — Type

Cylinder

Shape object of type cylinder.

Args

- radius : radius of cylinder
- thickness : thickness of the wall of cylinder
- length : length of cylinder

Returns

- obj : Shape object **Cylinder**.

PDMesh.Disk — Type

Disk

Shape object of type disk.

Args

- radius : radius of disk
- thickness : length of disk

Returns

- obj : Shape object **disk**.

PDMesh.PostOpObj — Type

PostOpObj

Create a post operation object (lazy) from object ( or post operation object) and an operation. Operations will be applied while a create call.

Returns

- obj : Post operation object (lazy)

PDMesh.Shape — Type

Shape
Abstarct shape object.

PDMesh.Sphere — Type

Sphere

Shape object of type sphere.

Args

- radius : radius of sphere

Returns

- obj : Shape object **sphere**.

PDMesh.changetype — Method

changetype(obj::T, f::Function, ntype::Int64) where T

Change mesh particle type for object using boolean array from function f.

PDMesh.create — Method

create(shape::T; resolution=nothing, rand_=0.0, type::Int64=1) where T <: Shape

Abstact function for creating Shape objects.

Returns

- X : Initial reference position 
- V : Initial velocity 
- Y : Initial position 
- volume : Volume per particle point 
- type : Type of particle point

PDMesh.delete — Method

delete(obj::T, f::Function) where T

Delete mesh particles for object using boolean array from function f.

PDMesh.move — Method

move(obj::T; by=[0.0, 0.0, 0.0]) where T

Move mesh particles for object by given "by".

PDMesh.rotate — Method

rotate(obj::T; angle=0.0, point=[0.0, 0.0, 0.0], vector_=[1.0, 0.0, 0.0]) where T <: Union{Shape, PostOpObj}

Rotate shape object by given angle about given vector and point.

PDMesh.velocity — Method

velocity(obj::T, f::Function, vel) where T

Change velocity of particles for object using boolean array from function f.